Refactor the metrics-dashboard component
