+++ dashboard/src/app/_components/metrics/stats-row.tsx
@@ -1,0 +1,59 @@
+"use client";
+
+import { Activity, Clock, Database, Sparkles, User } from "lucide-react";
+import { StatCard } from "./stat-card";
+import { type MetricsResponse } from "./types";
+
+interface StatsRowProps {
+  metrics: MetricsResponse;
+}
+
+export function StatsRow({ metrics }: Readonly<StatsRowProps>) {
+  const userActivity = metrics.user_activity;
+  const metricsData = metrics.metrics;
+
+  const uniqueUsers = userActivity.unique_users.toLocaleString("es-ES");
+  const totalEvents = userActivity.total_events.toLocaleString("es-ES");
+  const avgSession = userActivity.mean_session_length_seconds
+    ? (userActivity.mean_session_length_seconds / 60).toFixed(1)
+    : "0.0";
+  const ragLatency = metricsData.response_time
+    ? metricsData.response_time.toFixed(0)
+    : "0";
+  const totalMetrics = metricsData.total_count.toLocaleString("es-ES");
+
+  return (
+    <div className="mb-8 grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-5">
+      <StatCard
+        label="Usuarios únicos"
+        value={uniqueUsers}
+        helper="En el período"
+        icon={User}
+      />
+      <StatCard
+        label="Eventos totales"
+        value={totalEvents}
+        helper="Interacciones"
+        icon={Activity}
+      />
+      <StatCard
+        label="Sesión media"
+        value={`${avgSession}m`}
+        helper="Tiempo de uso"
+        icon={Clock}
+      />
+      <StatCard
+        label="Latencia RAG"
+        value={`${ragLatency}ms`}
+        helper="Respuesta LLM"
+        icon={Sparkles}
+      />
+      <StatCard
+        label="Métricas RAG"
+        value={totalMetrics}
+        helper="Registros totales"
+        icon={Database}
+      />
+    </div>
+  );
+}
