+++ dashboard/src/app/_components/metrics/utils.ts
@@ -1,0 +1,143 @@
+import { type MetricsResponse } from "./types";
+
+export const getDateFormatter = () =>
+  new Intl.DateTimeFormat("es-ES", {
+    dateStyle: "medium",
+    timeStyle: "short",
+  });
+
+export function isEmptyData(data: MetricsResponse): boolean {
+  return (
+    data?.user_activity?.total_events === 0 &&
+    data?.user_activity?.unique_users === 0
+  );
+}
+
+export function getErrorTitle(error: unknown): string {
+  if (!error) return "Error Loading Metrics";
+
+  const errorMessage =
+    error instanceof Error
+      ? error.message
+      : typeof error === "string"
+        ? error
+        : "Unknown error";
+
+  if (
+    errorMessage.includes("UNAUTHORIZED") ||
+    errorMessage.includes("You must be logged in")
+  ) {
+    return "Unauthorized Access";
+  }
+
+  if (
+    errorMessage.includes("FORBIDDEN") ||
+    errorMessage.includes("You must be an administrator") ||
+    errorMessage.includes("You do not have permission")
+  ) {
+    return "Access Denied";
+  }
+
+  if (errorMessage.includes("NOT_FOUND")) {
+    return "Data Not Found";
+  }
+
+  if (
+    errorMessage.includes("TIMEOUT") ||
+    errorMessage.includes("took too long")
+  ) {
+    return "Request Timeout";
+  }
+
+  if (
+    errorMessage.includes("INTERNAL_SERVER_ERROR") ||
+    errorMessage.includes("Failed to fetch") ||
+    errorMessage.includes("Failed to calculate")
+  ) {
+    return "Server Error";
+  }
+
+  return "Error Loading Metrics";
+}
+
+export function getErrorMessage(error: unknown): string {
+  if (!error)
+    return "No se pudieron recuperar los datos. Por favor, intenta nuevamente.";
+
+  const errorMessage =
+    error instanceof Error
+      ? error.message
+      : typeof error === "string"
+        ? error
+        : "Unknown error";
+
+  if (
+    errorMessage.includes("UNAUTHORIZED") ||
+    errorMessage.includes("You must be logged in")
+  ) {
+    return "No tienes autorización para acceder a estos datos. Por favor, inicia sesión nuevamente.";
+  }
+
+  if (
+    errorMessage.includes("FORBIDDEN") ||
+    errorMessage.includes("You must be an administrator") ||
+    errorMessage.includes("You do not have permission")
+  ) {
+    return "No tienes permisos para acceder a este recurso. Contacta a tu administrador si crees que esto es un error.";
+  }
+
+  if (errorMessage.includes("NOT_FOUND")) {
+    return errorMessage;
+  }
+
+  if (
+    errorMessage.includes("TIMEOUT") ||
+    errorMessage.includes("took too long")
+  ) {
+    return "La solicitud tardó demasiado en completarse. Por favor, intenta nuevamente.";
+  }
+
+  if (
+    errorMessage.includes("fetch failed") ||
+    errorMessage.includes("Network request failed")
+  ) {
+    return "Error de conexión. Por favor, verifica tu conexión a internet e intenta nuevamente.";
+  }
+
+  if (
+    errorMessage.includes("INTERNAL_SERVER_ERROR") ||
+    errorMessage.includes("Failed to fetch") ||
+    errorMessage.includes("Failed to calculate")
+  ) {
+    return "Error del servidor. Por favor, intenta nuevamente más tarde o contacta al soporte si el problema persiste.";
+  }
+
+  if (error instanceof Error && error.message) {
+    return error.message;
+  }
+
+  return "No se pudieron recuperar los datos. Por favor, intenta nuevamente.";
+}
+
+export function isRecoverableError(error: unknown): boolean {
+  if (!error) return true;
+
+  const errorMessage =
+    error instanceof Error
+      ? error.message
+      : typeof error === "string"
+        ? error
+        : "Unknown error";
+
+  if (
+    errorMessage.includes("UNAUTHORIZED") ||
+    errorMessage.includes("FORBIDDEN") ||
+    errorMessage.includes("You must be logged in") ||
+    errorMessage.includes("You must be an administrator") ||
+    errorMessage.includes("You do not have permission")
+  ) {
+    return false;
+  }
+
+  return true;
+}
